[store] refactored FilesystemAttachmentStore to use FileSaveTransactionRunnable and added tests for FilesystemTransactionStore.
git-svn-id: https://svn.xwiki.org/svnroot/xwiki/contrib/sandbox@33837 f329d543-caf0-0310-9063-dda96c69346f

+++ xwiki-store/xwiki-store-api/src/main/java/org/xwiki/store/FileSaveTransactionRunnable.java
@@ -108,6 +108,12 @@ public class FileSaveTransactionRunnable extends TransactionRunnable
      */
     protected void run() throws Exception
     {
+        if (!this.toSave.getParentFile().exists() && !this.toSave.getParentFile().mkdirs()) {
+            throw new IOException("Could not make directory tree to place file in. "
+                                  + "Do you have permission to write to ["
+                                  + this.toSave.getAbsolutePath() + "] ?");
+        }
+
         final InputStream in = this.provider.getStream();
         try {
             final OutputStream out = new FileOutputStream(this.tempFile);
